Atualizacao dos testes de integracao

+++ src/main/java/br/com/ubirataoliveira/desafio_todo_list/service/TodoService.java
@@ -2,12 +2,13 @@ package br.com.ubirataoliveira.desafio_todo_list.service;
 
 import java.util.List;
 
+import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.com.ubirataoliveira.desafio_todo_list.entity.Todo;
 import br.com.ubirataoliveira.desafio_todo_list.repository.TodoRepository;
-
-import org.springframework.data.domain.Sort;
 
 @Service
 public class TodoService {
@@ -33,12 +34,24 @@ public class TodoService {
         }
         
     public List<Todo> update(Todo todo){
-        todoRepository.save(todo);
-            
-        return list();
+    if (todo.getId() == null) {
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID é obrigatório");
+    }
+    
+    if (!todoRepository.existsById(todo.getId())) {
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo não encontrado");
+    }
+    
+    todoRepository.save(todo);
+    return list();
     }
 
+
     public List<Todo> delete(Long id){
+        if (!todoRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo não encontrado");
+        }
+        
         todoRepository.deleteById(id);
         return list();
     }
